+++ mayan/apps/tags/apps.py
@@ -132,11 +132,12 @@ class TagsApp(MayanAppConfig):
         SourceColumn(
             attribute='get_preview_widget', include_label=True, source=Tag
         )
-        SourceColumn(
+        source_column_tag_document_count = SourceColumn(
             func=lambda context: context['object'].get_document_count(
                 user=context['request'].user
             ), include_label=True, label=_('Documents'), source=Tag
         )
+        source_column_tag_document_count.add_exclude(source=DocumentTag)
 
         menu_facet.bind_links(
             links=(link_document_tag_list,), sources=(Document,)