+++ mayan/apps/tags/tests/test_views.py
@@ -3,6 +3,7 @@ from django.utils.encoding import force_text
 from mayan.apps.documents.tests.base import GenericDocumentViewTestCase
 from mayan.apps.testing.tests.base import GenericViewTestCase
 
+from ..links import link_tag_edit
 from ..models import Tag
 from ..permissions import (
     permission_tag_attach, permission_tag_create, permission_tag_delete,
@@ -57,10 +58,38 @@ class DocumentTagViewTestCase(
 
         response = self._request_test_document_tag_list_view()
         self.assertContains(
-            response=response, text=force_text(s=self.test_tag), status_code=200
+            response=response, text=force_text(s=self.test_tag),
+            status_code=200
         )
 
-    def test_document_attach_tag_view_no_permission(self):
+    def test_document_tags_list_tag_edit_link_with_full_access(self):
+        # Ensure that DocumentTag instances and links are
+        # resolved in this view and not base Tag instances.
+        self._create_test_tag(add_test_document=True)
+
+        self.grant_access(
+            obj=self.test_document, permission=permission_tag_view
+        )
+        self.grant_access(
+            obj=self.test_tag, permission=permission_tag_view
+        )
+        self.grant_access(
+            obj=self.test_tag, permission=permission_tag_edit
+        )
+
+        response = self._request_test_document_tag_list_view()
+
+        link_context = response.context[-1]
+        link_context['object'] = self.test_tag
+
+        result = link_tag_edit.resolve(context=link_context)
+
+        self.assertNotContains(
+            response=response, text=result.url,
+            status_code=200
+        )
+
+    def test_document_tag_attach_view_no_permission(self):
         self._create_test_tag()
 
         response = self._request_test_document_tag_attach_view()
@@ -68,7 +97,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
-    def test_document_attach_tag_view_with_tag_access(self):
+    def test_document_tag_attach_view_with_tag_access(self):
         self._create_test_tag()
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_attach)
@@ -78,7 +107,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
-    def test_document_attach_tag_view_with_document_access(self):
+    def test_document_tag_attach_view_with_document_access(self):
         self._create_test_tag()
 
         self.grant_access(
@@ -90,7 +119,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
-    def test_document_attach_tag_view_with_full_access(self):
+    def test_document_tag_attach_view_with_full_access(self):
         self._create_test_tag()
 
         self.grant_access(
@@ -103,7 +132,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag in self.test_document.tags.all())
 
-    def test_document_multiple_attach_tag_view_no_permission(self):
+    def test_document_multiple_tag_attach_view_no_permission(self):
         self._create_test_tag()
 
         response = self._request_test_document_multiple_tag_multiple_attach_view()
@@ -111,7 +140,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
-    def test_document_multiple_attach_tag_view_with_tag_access(self):
+    def test_document_multiple_tag_attach_view_with_tag_access(self):
         self._create_test_tag()
 
         self.grant_access(obj=self.test_tag, permission=permission_tag_attach)
@@ -121,7 +150,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
-    def test_document_multiple_attach_tag_view_with_document_access(self):
+    def test_document_multiple_tag_attach_view_with_document_access(self):
         self._create_test_tag()
 
         self.grant_access(
@@ -133,7 +162,7 @@ class DocumentTagViewTestCase(
 
         self.assertTrue(self.test_tag not in self.test_document.tags.all())
 
-    def test_document_multiple_attach_tag_view_with_full_access(self):
+    def test_document_multiple_tag_attach_view_with_full_access(self):
         self._create_test_tag()
 
         self.grant_access(